# _reference/codebases/wa-blaster/docs/chatbot-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tickets/{id}": {
      "get": {
        "operationId": "TicketsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/tickets/{id}/agent-context": {
      "get": {
        "operationId": "TicketsController_agentContext",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/tickets/{id}/suggest-reply": {
      "post": {
        "operationId": "TicketsController_suggestReply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/tickets/{id}/assign": {
      "post": {
        "operationId": "TicketsController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/tickets/{id}/resolve": {
      "post": {
        "operationId": "TicketsController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/tickets/{id}/close": {
      "post": {
        "operationId": "TicketsController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/tickets/{id}/reopen": {
      "post": {
        "operationId": "TicketsController_reopen",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/tickets/{id}/knowledge-suggestion": {
      "get": {
        "operationId": "TicketsController_knowledgeSuggestion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/tickets/{id}/knowledge-candidate": {
      "post": {
        "operationId": "TicketsController_createKnowledgeCandidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKnowledgeCandidateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "WhatsApp Blasting API",
    "description": "REST API for the blasting platform and the AI chatbot inbox + admin surface.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AssignTicketDto": {
        "type": "object",
        "properties": {}
      },
      "CreateKnowledgeCandidateDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}